feat(editable): update for display panel modal
