+++ packages/theme/lib/junipero.sass
@@ -16,3 +16,4 @@
   @import "@junipero/theme/dist/css/ColorField.min"
   @import "@junipero/theme/dist/css/TouchableZone.min"
   @import "@junipero/theme/dist/css/Card.min"
+  @import "@junipero/theme/dist/css/Modal.min"
